requirements.txt
Package                          Version
-------------------------------- ----------
Adafruit-Blinka                  5.4.1
adafruit-circuitpython-busdevice 5.0.1
adafruit-circuitpython-si7021    3.2.1
Adafruit-PlatformDetect          2.17.0
Adafruit-PureIO                  1.1.6
aiohttp                          3.6.2
async-timeout                    3.0.1
attrs                            19.3.0
certifi                          2019.11.28
cffi                             1.13.2
chardet                          3.0.4
colorzero                        1.1
cryptography                     2.8
cycler                           0.10.0
Cython                           0.29.14
future                           0.18.2
gpiozero                         1.5.1
idna                             2.8
idna-ssl                         1.1.0
kiwisolver                       1.1.0
MCP342x                          0.3.4
multidict                        4.7.4
mysql                            0.0.2
mysqlclient                      1.4.4
numpy                            1.18.1
pip                              20.0.2
plotly                           4.5.0
pycparser                        2.19
pyftdi                           0.52.0
pyngrok                          2.0.3
pyparsing                        2.4.6
pyserial                         3.4
python-dateutil                  2.8.1
python-telegram-bot              12.3.0
pytz                             2019.3
pyusb                            1.1.0
PyYAML                           5.3
retrying                         1.3.3
rpi-ws281x                       4.2.4
RPi.bme280                       0.2.3
RPi.GPIO                         0.7.0
setuptools                       28.8.0
six                              1.13.0
smbus                            1.1.post2
smbus2                           0.3.0
spidev                           3.4
sysv-ipc                         1.0.1
telepot                          12.7
tornado                          6.0.3
typing-extensions                3.7.4.1
urllib3                          1.25.7
yarl                             1.4.2
